1. Read data from file sync without importing json that will increase odule initiation and parsing and may help in iOS issue 2. misc typescript related changes 3. version bump from v3.0.4 to v3.0.5
