+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "country-state-city",
-  "version": "3.0.4",
+  "version": "3.0.5",
   "description": "Library for fetching Country, its States and Cities",
   "main": "./lib/cjs/index.js",
   "module": "./lib/index.js",
