+++ tsconfig.json
@@ -10,8 +10,14 @@
     "moduleResolution": "node",
     "resolveJsonModule": true
   },
+  "ts-node": {
+    // these options are overrides used only by ts-node
+    "compilerOptions": {
+      "module": "commonjs"
+    }
+  },
   "include": [
-    "./src"
+    "./src/**/*"
   ],
   "exclude": [
     "src/__test__"
